Move test scripts to separate directory
