+++ __tests__/reader.int.js
@@ -1,4 +1,4 @@
-const Reader = require("./reader").Reader;
+const Reader = require("../lib/reader").Reader;
 
 const bufferPrefix = Buffer.alloc(4, 0x00);
